requirements.txt
fastapi
uvicorn
python-dotenv
google-generativeai
langdetect
gTTS
edge-tts
pydantic